requirements.txt
openai-whisper>=20231117
edge-tts>=6.1.9
pydub>=0.25.1
requests>=2.31.0
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
PyAudio>=0.2.13
python-dotenv>=1.0.0

